Add files via upload
A project in the Java programming language. The idea of the project is to develop a testing simulator for a bus station manager, which includes multiple stations, including a refueling station, entrance and exit gates, and numerous staff members. The project involves parallel programming, reading and writing data from files, and other complex programming elements.